Safari-609.1.11.1/LayoutTests/accessibility/mac/caret-browsing-tab-selection-expected.txt

foo
A paragraph with a single line

bar
This tests caret browsing for html select elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS webArea.role is 'AXRole: AXWebArea'
PASS caretBrowsingEnabled(webArea) is false
PASS elementAtStartMarkerOfSelectedTextMarkerRange(webArea).stringValue is 'AXValue: foo'
PASS elementAtStartMarkerOfSelectedTextMarkerRange(webArea).stringValue is 'AXValue: foo'
PASS accessibilityController.enhancedAccessibilityEnabled is true
PASS caretBrowsingEnabled(webArea) is true
PASS elementAtStartMarkerOfSelectedTextMarkerRange(webArea).stringValue is 'AXValue: foo'
PASS elementAtStartMarkerOfSelectedTextMarkerRange(webArea).stringValue is 'AXValue: bar'
PASS successfullyParsed is true

TEST COMPLETE
